ci: disable docker image cache (#943)
* ci: disable docker image cache

The Docker image cache was invaluable when the ARM build was still emulated. Now that all builds are run on native hardware the build time has reduced significantly. As such the cache is doing more harm than it is useful. Leading to invalid Dockerfiles ending up on the default branch.

* chore: remove libc6-dev